Refactor media handling and inline flows
