+++ app/service/view/inline/__init__.py
@@ -4,6 +4,7 @@ from __future__ import annotations
 
 import logging
 from dataclasses import dataclass, field
+
 from navigator.core.entity.history import Message
 from navigator.core.service.rendering import decision as D
 from navigator.core.service.rendering.config import RenderingConfig
@@ -19,123 +20,96 @@ from ..executor import EditExecutor
 from ...store import preserve
 
 
-@dataclass(slots=True)
-class InlineHandler:
-    """Mediate inline edits, remapping fallbacks when required."""
+class InlineTelemetryAdapter:
+    """Emit inline-related telemetry events."""
 
-    guard: InlineGuard
-    remapper: InlineRemapper
-    editor: InlineEditor
-    telemetry: Telemetry
-    _channel: TelemetryChannel = field(init=False, repr=False)
+    def __init__(self, telemetry: Telemetry) -> None:
+        self._channel: TelemetryChannel = telemetry.channel(__name__)
 
-    def __post_init__(self) -> None:
-        self._channel = self.telemetry.channel(__name__)
+    def remap_delete_send(self, target: D.Decision) -> None:
+        self._channel.emit(
+            logging.INFO,
+            LogCode.INLINE_REMAP_DELETE_SEND,
+            origin="DELETE_SEND",
+            target=target.name,
+        )
 
-    async def handle(
-            self,
-            *,
-            scope: Scope,
-            payload: Payload,
-            tail: Message | None,
-            executor: EditExecutor,
-            config: RenderingConfig,
-    ) -> InlineOutcome | None:
-        """Plan inline reconciliation for the provided ``payload``."""
+    def deny_switch(self) -> None:
+        self._channel.emit(logging.INFO, LogCode.INLINE_CONTENT_SWITCH_FORBIDDEN)
 
-        entry = self._prepare_candidate(payload, tail)
-        base = tail
-        media = getattr(entry, "media", None)
 
-        if media:
-            if not self.guard.admissible(entry):
-                return await self._fallback(scope, entry, base, executor)
-            verdict = D.decide(tail, entry, config)
-            return await self._mediate(scope, verdict, entry, base, executor)
+class InlineCandidateBuilder:
+    """Prepare preserved payloads suitable for inline editing."""
 
-        return await self._scribe(scope, entry, base, executor)
-
-    def _prepare_candidate(self, payload: Payload, tail: Message | None) -> Payload:
-        """Derive preserved inline payload ready for inline editing."""
-
+    def create(self, payload: Payload, tail: Message | None) -> Payload:
         entry = preserve(payload, tail)
         if getattr(entry, "group", None):
             return entry.morph(media=entry.group[0], group=None)
         return entry
 
-    async def _mediate(
-            self,
-            scope: Scope,
-            verdict: D.Decision,
-            entry: Payload,
-            base: Message | None,
-            executor: EditExecutor,
+
+class InlineEditMediator:
+    """Delegate inline edits to the executor with consistent wiring."""
+
+    def __init__(self, editor: InlineEditor) -> None:
+        self._editor = editor
+
+    async def apply(
+        self,
+        scope: Scope,
+        verdict: D.Decision,
+        payload: Payload,
+        base: Message | None,
+        *,
+        executor: EditExecutor,
     ) -> InlineOutcome | None:
-        """Apply verdict-specific inline reconciliation logic."""
-
-        if verdict is D.Decision.DELETE_SEND:
-            return await self._handle_delete_send(scope, entry, base, executor=executor)
-
-        if verdict in (D.Decision.EDIT_MEDIA_CAPTION, D.Decision.EDIT_MARKUP):
-            return await self._apply_edit(scope, verdict, entry, base, executor=executor)
-
-        return await self._apply_edit(
+        return await self._editor.apply(
             scope,
-            D.Decision.EDIT_MEDIA,
-            entry,
+            verdict,
+            payload,
             base,
             executor=executor,
         )
 
-    async def _scribe(
-            self,
-            scope: Scope,
-            entry: Payload,
-            base: Message | None,
-            executor: EditExecutor,
+
+class InlineMediaFlow:
+    """Coordinate inline media decisions and fallbacks."""
+
+    def __init__(
+        self,
+        guard: InlineGuard,
+        remapper: InlineRemapper,
+        mediator: InlineEditMediator,
+        telemetry: InlineTelemetryAdapter,
+    ) -> None:
+        self._guard = guard
+        self._remapper = remapper
+        self._mediator = mediator
+        self._telemetry = telemetry
+
+    async def handle(
+        self,
+        scope: Scope,
+        entry: Payload,
+        base: Message | None,
+        executor: EditExecutor,
+        config: RenderingConfig,
     ) -> InlineOutcome | None:
-        """Handle inline updates when the preserved payload is textual."""
-
-        if base and getattr(base, "media", None):
-            adjusted = entry.morph(media=base.media, group=None)
-            if self._requires_caption_refresh(adjusted):
-                return await self._apply_edit(
-                    scope,
-                    D.Decision.EDIT_MEDIA_CAPTION,
-                    adjusted,
-                    base,
-                    executor=executor,
-                )
-            if not match(getattr(base, "markup", None), getattr(entry, "reply", None)):
-                return await self._apply_edit(
-                    scope,
-                    D.Decision.EDIT_MARKUP,
-                    adjusted,
-                    base,
-                    executor=executor,
-                )
-            return self._deny_switch()
-
-        return await self._apply_edit(
-            scope,
-            D.Decision.EDIT_TEXT,
-            entry,
-            base,
-            executor=executor,
-        )
+        if not self._guard.admissible(entry):
+            return await self._fallback(scope, entry, base, executor)
+        verdict = D.decide(base, entry, config)
+        return await self._mediate(scope, verdict, entry, base, executor)
 
     async def _fallback(
-            self,
-            scope: Scope,
-            entry: Payload,
-            base: Message | None,
-            executor: EditExecutor,
+        self,
+        scope: Scope,
+        entry: Payload,
+        base: Message | None,
+        executor: EditExecutor,
     ) -> InlineOutcome | None:
-        """Attempt graceful inline fallback when guard rejects the payload."""
-
         if base and not match(getattr(base, "markup", None), getattr(entry, "reply", None)):
             adjusted = entry.morph(media=base.media if base else None, group=None)
-            outcome = await self._apply_edit(
+            outcome = await self._mediator.apply(
                 scope,
                 D.Decision.EDIT_MARKUP,
                 adjusted,
@@ -144,45 +118,48 @@ class InlineHandler:
             )
             if outcome:
                 return outcome
-        return self._deny_switch()
+        self._telemetry.deny_switch()
+        return None
 
-    async def _apply_edit(
-            self,
-            scope: Scope,
-            verdict: D.Decision,
-            payload: Payload,
-            base: Message | None,
-            executor: EditExecutor,
+    async def _mediate(
+        self,
+        scope: Scope,
+        verdict: D.Decision,
+        entry: Payload,
+        base: Message | None,
+        executor: EditExecutor,
     ) -> InlineOutcome | None:
-        """Delegate inline edits to the executor with consistent wiring."""
-
-        return await self.editor.apply(
+        if verdict is D.Decision.DELETE_SEND:
+            return await self._handle_delete_send(scope, entry, base, executor=executor)
+        if verdict in (D.Decision.EDIT_MEDIA_CAPTION, D.Decision.EDIT_MARKUP):
+            return await self._mediator.apply(
+                scope,
+                verdict,
+                entry,
+                base,
+                executor=executor,
+            )
+        return await self._mediator.apply(
             scope,
-            verdict,
-            payload,
+            D.Decision.EDIT_MEDIA,
+            entry,
             base,
             executor=executor,
         )
 
     async def _handle_delete_send(
-            self,
-            scope: Scope,
-            entry: Payload,
-            base: Message | None,
-            executor: EditExecutor,
+        self,
+        scope: Scope,
+        entry: Payload,
+        base: Message | None,
+        *,
+        executor: EditExecutor,
     ) -> InlineOutcome | None:
-        """Resolve delete-send verdicts through inline remapping."""
-
-        remapped = self.remapper.remap(base, entry)
-        self._channel.emit(
-            logging.INFO,
-            LogCode.INLINE_REMAP_DELETE_SEND,
-            origin="DELETE_SEND",
-            target=remapped.name,
-        )
+        remapped = self._remapper.remap(base, entry)
+        self._telemetry.remap_delete_send(remapped)
         if remapped is D.Decision.EDIT_MARKUP:
             adjusted = entry.morph(media=base.media if base else None, group=None)
-            return await self._apply_edit(
+            return await self._mediator.apply(
                 scope,
                 remapped,
                 adjusted,
@@ -190,7 +167,7 @@ class InlineHandler:
                 executor=executor,
             )
         if remapped in (D.Decision.EDIT_TEXT, D.Decision.EDIT_MEDIA):
-            outcome = await self._apply_edit(
+            outcome = await self._mediator.apply(
                 scope,
                 remapped,
                 entry,
@@ -199,17 +176,57 @@ class InlineHandler:
             )
             if outcome:
                 return outcome
-        return self._deny_switch()
-
-    def _deny_switch(self) -> None:
-        """Emit telemetry for forbidden inline content switches."""
-
-        self._channel.emit(logging.INFO, LogCode.INLINE_CONTENT_SWITCH_FORBIDDEN)
+        self._telemetry.deny_switch()
         return None
 
-    def _requires_caption_refresh(self, payload: Payload) -> bool:
-        """Report whether inline payload changes caption semantics."""
 
+class InlineTextFlow:
+    """Handle inline edits when payloads are textual."""
+
+    def __init__(
+        self,
+        mediator: InlineEditMediator,
+        telemetry: InlineTelemetryAdapter,
+    ) -> None:
+        self._mediator = mediator
+        self._telemetry = telemetry
+
+    async def handle(
+        self,
+        scope: Scope,
+        entry: Payload,
+        base: Message | None,
+        executor: EditExecutor,
+    ) -> InlineOutcome | None:
+        if base and getattr(base, "media", None):
+            adjusted = entry.morph(media=base.media, group=None)
+            if self._requires_caption_refresh(adjusted):
+                return await self._mediator.apply(
+                    scope,
+                    D.Decision.EDIT_MEDIA_CAPTION,
+                    adjusted,
+                    base,
+                    executor=executor,
+                )
+            if not match(getattr(base, "markup", None), getattr(entry, "reply", None)):
+                return await self._mediator.apply(
+                    scope,
+                    D.Decision.EDIT_MARKUP,
+                    adjusted,
+                    base,
+                    executor=executor,
+                )
+            self._telemetry.deny_switch()
+            return None
+        return await self._mediator.apply(
+            scope,
+            D.Decision.EDIT_TEXT,
+            entry,
+            base,
+            executor=executor,
+        )
+
+    def _requires_caption_refresh(self, payload: Payload) -> bool:
         extra = payload.extra or {}
         return bool(
             payload.text is not None
@@ -219,4 +236,55 @@ class InlineHandler:
         )
 
 
-__all__ = ["InlineHandler", "InlineOutcome", "InlineEditor", "InlineGuard", "InlineRemapper"]
+@dataclass(slots=True)
+class InlineHandler:
+    """Mediate inline edits, remapping fallbacks when required."""
+
+    guard: InlineGuard
+    remapper: InlineRemapper
+    editor: InlineEditor
+    telemetry: Telemetry
+    _candidate: InlineCandidateBuilder = field(init=False, repr=False)
+    _media: InlineMediaFlow = field(init=False, repr=False)
+    _text: InlineTextFlow = field(init=False, repr=False)
+
+    def __post_init__(self) -> None:
+        adapter = InlineTelemetryAdapter(self.telemetry)
+        mediator = InlineEditMediator(self.editor)
+        self._candidate = InlineCandidateBuilder()
+        self._media = InlineMediaFlow(self.guard, self.remapper, mediator, adapter)
+        self._text = InlineTextFlow(mediator, adapter)
+
+    async def handle(
+        self,
+        *,
+        scope: Scope,
+        payload: Payload,
+        tail: Message | None,
+        executor: EditExecutor,
+        config: RenderingConfig,
+    ) -> InlineOutcome | None:
+        """Plan inline reconciliation for the provided ``payload``."""
+
+        entry = self._candidate.create(payload, tail)
+        media = getattr(entry, "media", None)
+
+        if media:
+            return await self._media.handle(
+                scope,
+                entry,
+                tail,
+                executor,
+                config,
+            )
+
+        return await self._text.handle(scope, entry, tail, executor)
+
+
+__all__ = [
+    "InlineHandler",
+    "InlineOutcome",
+    "InlineEditor",
+    "InlineGuard",
+    "InlineRemapper",
+]
